+++ src/C++/ThreadedSocketInitiator.cpp
@@ -60,14 +60,14 @@ ThreadedSocketInitiator::~ThreadedSocketInitiator()
 void ThreadedSocketInitiator::onConfigure( const SessionSettings& s )
 throw ( ConfigError )
 {
-  try { m_reconnectInterval = s.get().getLong( "ReconnectInterval" ); }
+  try { m_reconnectInterval = s.get().getInt( "ReconnectInterval" ); }
   catch ( std::exception& ) {}
   if( s.get().has( SOCKET_NODELAY ) )
     m_noDelay = s.get().getBool( SOCKET_NODELAY );
   if( s.get().has( SOCKET_SEND_BUFFER_SIZE ) )
-    m_sendBufSize = s.get().getLong( SOCKET_SEND_BUFFER_SIZE );
+    m_sendBufSize = s.get().getInt( SOCKET_SEND_BUFFER_SIZE );
   if( s.get().has( SOCKET_RECEIVE_BUFFER_SIZE ) )
-    m_rcvBufSize = s.get().getLong( SOCKET_RECEIVE_BUFFER_SIZE );
+    m_rcvBufSize = s.get().getInt( SOCKET_RECEIVE_BUFFER_SIZE );
 }
 
 void ThreadedSocketInitiator::onInitialize( const SessionSettings& s )
@@ -251,13 +251,13 @@ void ThreadedSocketInitiator::getHost( const SessionID& s, const Dictionary& d,
   if( d.has(hostString) && d.has(portString) )
   {
     address = d.getString( hostString );
-    port = ( short ) d.getLong( portString );
+    port = ( short ) d.getInt( portString );
   }
   else
   {
     num = 0;
     address = d.getString( SOCKET_CONNECT_HOST );
-    port = ( short ) d.getLong( SOCKET_CONNECT_PORT );
+    port = ( short ) d.getInt( SOCKET_CONNECT_PORT );
   }
 
   m_sessionToHostNum[ s ] = ++num;
