+++ src/C++/test/DictionaryTestCase.cpp
@@ -47,14 +47,14 @@ TEST(setGetString)
 TEST(setGetLong)
 {
   Dictionary object;
-  object.setLong( "LONGKEY1", 12 );
-  object.setLong( "LONGKEY2", 9827362 );
+  object.setInt( "LONGKEY1", 12 );
+  object.setInt( "LONGKEY2", 9827362 );
   object.setString( "BADLONGKEY", "AB12" );
 
-  CHECK_EQUAL( 12, object.getLong( "LONGKEY1" ) );
-  CHECK_EQUAL( 9827362, object.getLong( "LONGKEY2" ) );
-  CHECK_THROW( object.getLong( "LONGKEY3" ), ConfigError );
-  CHECK_THROW( object.getLong( "BADLONGKEY" ), ConfigError );
+  CHECK_EQUAL( 12, object.getInt( "LONGKEY1" ) );
+  CHECK_EQUAL( 9827362, object.getInt( "LONGKEY2" ) );
+  CHECK_THROW( object.getInt( "LONGKEY3" ), ConfigError );
+  CHECK_THROW( object.getInt( "BADLONGKEY" ), ConfigError );
 }
 
 TEST(setGetDouble)
